requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.2.0
black==22.3.0
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
discord.py==1.7.3
distlib==0.3.4
filelock==3.7.0
flake8==4.0.1
frozenlist==1.3.0
identify==2.5.0
idna==3.3
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.19.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyNaCl==1.4.0
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.0.1
virtualenv==20.14.1
yarl==1.7.2
